[SYCL][NFC] Sink codeLocationCallback to ur.cpp
The codeLocationCallback function is only used in ur.cpp, so there is no
need to have it in common.hpp/cpp.
